renamed export files to index.js
